Added and applied auth policies

+++ NetBlog.API/NetBlog.API/Controllers/CommentsController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using NetBlog.BAL.Services.CommentsService;
 using NetBlog.Common.DTO;
 
@@ -9,13 +10,16 @@ namespace NetBlog.API.Controllers
     public class CommentsController : Controller
     {
         private readonly ICommentsService _commentService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public CommentsController(ICommentsService commentService)
+        public CommentsController(ICommentsService commentService, IAuthorizationService authorizationService)
         {
             _commentService = commentService;
+            _authorizationService = authorizationService;
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> CreateComment(CreateCommentDTO dto)
         {
             var comment = await _commentService.CreateComment(dto);
@@ -25,11 +29,16 @@ namespace NetBlog.API.Controllers
         [HttpDelete("{commentId}")]
         public async Task<IActionResult> DeleteComment(Guid commentId)
         {
-            var comment = await _commentService.DeleteComment(commentId);
+            var tryAuth = await _authorizationService.AuthorizeAsync(User, commentId, "CanDeleteCommentPolicy");
+            if (tryAuth.Succeeded)
+            {
+                var comment = await _commentService.DeleteComment(commentId);
 
-            if(comment ==null) return NotFound("Comment not found");
+                if (comment == null) return NotFound("Comment not found");
 
-            return Ok(comment);
+                return Ok(comment);
+            }
+            return new ForbidResult();
         }
 
         [HttpGet("post/{postId}")]
